Refactor import paths and annotate API endpoints; Enhanced Swagger Docs
