+++ src/modules/team/interface/__test__/team.e2e.spec.ts
@@ -12,7 +12,7 @@ import { loadFixtures } from '@data/util/loader';
 import { AppModule } from '@/app.module';
 import { COGNITO_AUTH } from '@/common/cognito/application/interface/cognito.service.interface';
 import { SuccessResponseInterceptor } from '@/common/response_service/interceptor/success_response.interceptor';
-import { AUTH_RESPONSE } from '@/modules/authorization/infraestructure/policy/exceptions/auth-error';
+import { AUTH_RESPONSE } from '@/modules/authorization/exceptions/auth-error';
 import { identityProviderServiceMock } from '@/test/test.module.bootstrapper';
 import { createAccessToken, makeRequest } from '@/test/test.util';
 
@@ -183,9 +183,9 @@ describe('Team - [/team]', () => {
         data: {
           name: 'team updated',
           usersEmails: ['user1'],
+          id: 'team0',
         },
       });
-
       expect(response.body.payload).toEqual({
         name: 'team updated',
         id: 'team0',
